+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- HTTP API: Oauth2 secret tokens now accepts null refresh-token. Some oauth2 token providers can return un-expirable access_tokens. 
+
 ## [7.7.2] 2025-01-27
 
 ### Fixed